Moved around some of the test bits, shifted direction of the DWT pruning
