+++ Phase3ResNet/DWTResNetPruning.py
@@ -1,50 +1,24 @@
-from absl import app, flags
-import tensorflow as tf
-import numpy as np
 import os
+import json
+import tensorflow as tf
+from transformers import TFResNetModel
 import pywt
-import csv
-import ResNetRandomPruning as DWTR
-from transformers import TFResNetForImageClassification, ResNetConfig
+import numpy as np
+from absl import app, flags, logging
 
-"""
-TODO threshold values:
-0, 0.236, 0.382, 0.5, 0.618, 0.786, 1
-"""
-# Define Flags.
 FLAGS = flags.FLAGS
-'''
-- Wavelet: Type of wavelet to use for the Discrete Wavelet Transform (DWT).
-    The choice of wavelet affects the base wavelet function used in the DWT,
-    which can impact the model's ability to learn from data 
-    that has been decomposed in different ways.
-    Different wavelets capture different characteristics of the input data,
-    potentially influencing the features learned by the model.
-- Levels: Decomposition level for the DWT.
-    A higher level results in a deeper decomposition of the input data,
-        affecting the granularity of the wavelet transform applied to the input data.
-    Higher levels may capture more abstract features 
-        but also increase computational complexity and the risk of overfitting.
-- Threshold: Threshold for thresholding the wavelet coefficients obtained from the DWT.
-    A value of 0 means no thresholding is applied, 
-        while the maximum value corresponds to the absolute value 
-        of the largest wavelet coefficient. 
-    Thresholding controls the sparsity of the wavelet coefficients,
-    potentially impacting the interpretability of the learned features 
-        and the model's ability to generalize.
-'''
+
+# Define command-line flags
 flags.DEFINE_string('model_path', None,
-                    'Full path to the original DWT TensorFlow model.')
-flags.DEFINE_enum('wavelet', 'haar', ['haar', 'db1', 'db2', 'coif1', 'bior1.3', 'rbio1.3', 'sym2', 'mexh', 'morl'],
-                  'Type of wavelet to use for DWT.')
-flags.DEFINE_integer('level', '1', 'Decomposition level for the DWT.')
-flags.DEFINE_float('threshold', '1.0',
-                   'Threshold for thresholding the wavelet coefficients obtained from the DWT.')
-flags.DEFINE_string('csv_path', 'experiment_log.csv',
-                    'Path to the CSV log file.')
-flags.DEFINE_string('guid', None, 'GUID for the current pruning operation.')
-# flags.mark_flag_as_required('model_path')
-flags.mark_flag_as_required('guid')
+                    'Path to the pre-trained model to be pruned.')
+flags.DEFINE_float('threshold', None,
+                   'Threshold value for identifying insignificant weights.')
+flags.DEFINE_string('wavelet_type', None,
+                    'Type of wavelet function used in the decomposition.')
+flags.DEFINE_integer('decomp_level', None,
+                     'Decomposition level for wavelet transform.')
+flags.DEFINE_string(
+    'save_path', None, 'Directory path where the pruned model and logs are saved.')
 
 
 def setup_tensorflow_gpu():
@@ -58,330 +32,151 @@ def setup_tensorflow_gpu():
         except RuntimeError as e:
             print("Error setting up GPU:", e)
 
-# Load model
 
-
-def get_model(model_path):
+def save_and_log_results(model, pruned_weights_info, save_path, model_type):
     """
-    gets the model from the defined location.
-    """
-    # Load the pre-trained model
-    try:
-        model = tf.keras.models.load_model(model_path)
-        return model
-    except Exception as e:
-        print("Failed to load model:", e)
-        return None
-
-# Save model
-
-
-def save_model(model, original_model_path, guid, isRandomPruned=False):
-    """
-    Save the model in TensorFlow SavedModel format.
+    Saves the pruned model and logs detailed information about the pruning process to both a text file and a structured JSON file for easier analysis.
 
     Args:
-        model: The TensorFlow model to save.
-        original_model_path: The path to the original model, used to determine save directory.
-        guid: Unique identifier for the save directory.
-        isRandomPruned: Flag to indicate if the model was pruned randomly.
-    """
-    # Determine the directory of the original model
-    directory = os.path.dirname(
-        original_model_path) if original_model_path is not None else os.getcwd()
-
-    # Create a new directory name with the GUID
-    new_directory_name = f"{directory}/pruned_{guid}"
-    os.makedirs(new_directory_name, exist_ok=True)
-
-    # Determine the save path based on pruning type
-    model_type = "RP" if isRandomPruned else "DWT"
-    model_save_path = f"{new_directory_name}/model_{model_type}_threshold{FLAGS.threshold}"
-
-    try:
-        # Save the model using PretrainedSavedModel
-        model.save_pretrained(model_save_path)
-        print(f"Model saved successfully at {model_save_path}")
-    except Exception as e:
-        print(f"Failed to save the model: {e}")
-        raise
-
-
-def log_to_csv(guid, details):
-    """
-    Perfoms the logging
-    """
-    try:
-        csv_path = flags.FLAGS.csv_path
-        file_exists = os.path.isfile(csv_path)
-        with open(csv_path, mode='a', newline='') as file:
-            writer = setup_csv_writer(file, file_exists)
-            row_data = {'GUID': guid, **details}
-            writer.writerow(row_data)
-    except Exception as e:
-        print(f"Failed to log data to CSV: {e}")
-        raise
-
-
-def setup_csv_writer(file, file_exists):
-    """
-    executes the logging
-    """
-    fieldnames = [
-        'GUID', 'Wavelet', 'Level', 'Threshold', 'DWT Phase',
-        'Original Parameter Count', 'Prune count', 'Layers skipped', 'Total Pruned Count'
-    ]
-    writer = csv.DictWriter(file, fieldnames=fieldnames)
-    if not file_exists:
-        writer.writeheader()
-    return writer
-
-
-# DWT process
-
-# DWT all weights
-
-
-def optimize_model(threshold, wavelet, level, guid, model=None):
-    """
-    Optimizes the weights of a pre-trained model by applying wavelet-based pruning.
-
-    Args:
-        threshold (float): The threshold value used for pruning wavelet coefficients.
-        wavelet (str): The type of wavelet to be used for the wavelet transform.
-        level (int): The decomposition level for the wavelet transform.
-        guid (str): Unique identifier for the experiment.
-        model: The pre-trained model to be optimized.
+        model (tf.keras.Model): The pruned model to be saved.
+        pruned_weights_info (dict): Detailed information about the weights pruned, including count and percentage of total.
+        save_path (str): The directory path where the pruned model and logs are saved.
+        model_type (str): Identifier for the type of pruning ('DWT' or 'Random') to distinguish between save files.
 
     Returns:
-        int: The total number of pruned parameters.
+        str: The path to the saved model file.
     """
-    try:
-        original_param_count = model.count_params()
-        skipped_layers = 0
-        pruned_counts = []
+    # Create the save directory if it doesn't exist
+    os.makedirs(save_path, exist_ok=True)
 
-        # Collect all Conv2D layers
-        conv2d_layers = [layer for layer in model.layers if isinstance(
-            layer, tf.keras.layers.Conv2D)]
+    # Save the pruned model
+    model_save_path = os.path.join(save_path, f"pruned_model_{model_type}.h5")
+    model.save(model_save_path)
 
-        for layer in conv2d_layers:
-            pruned_count_temp = process_layer(layer, threshold, wavelet, level)
-            pruned_counts.append(pruned_count_temp)
+    # Log the pruned weights information to a text file
+    log_file_path = os.path.join(save_path, f"pruning_log_{model_type}.txt")
+    with open(log_file_path, "w") as log_file:
+        for layer_name, layer_info in pruned_weights_info.items():
+            log_file.write(f"Layer: {layer_name}\n")
+            log_file.write(
+                f"  Pruned weights count: {layer_info['pruned_weights_count']}\n")
+            log_file.write(
+                f"  Pruned weights percentage: {layer_info['pruned_weights_percentage']:.2f}%\n")
+            log_file.write("\n")
 
-        # Handle layers without weights
-        skipped_layers = handle_non_conv2d_layers(model)
+    # Log the pruned weights information to a JSON file
+    json_file_path = os.path.join(save_path, f"pruning_info_{model_type}.json")
+    with open(json_file_path, "w") as json_file:
+        json.dump(pruned_weights_info, json_file, indent=4)
 
-        total_pruned_count = sum(pruned_counts)
-        print(f"Total parameters pruned: {total_pruned_count}")
-
-        # Save the optimized model
-        save_model(model, original_model_path=FLAGS.model_path, guid=guid)
-
-        # Log model changes
-        log_model_changes(wavelet, level, threshold, skipped_layers,
-                          original_param_count, total_pruned_count, guid)
-    except Exception as e:
-        print(f"An error occurred during model optimization: {e}")
-        raise
-
-    return total_pruned_count
+    return model_save_path
 
 
-def process_layer(layer, threshold, wavelet, level, total_pruned_count):
+def prune_layer_tf(layer, threshold, wavelet_type, decomp_level):
     """
-    This is looking like many changes will be coming soon?
-    """
-    print(f"Processing layer: {layer.name}")
-    weights = layer.get_weights()[0]
-    print("Getting Coeffecients and Original shape")
-    coeffs, original_shape = apply_dwt(
-        weights=weights, wavelet=wavelet, level=level)
-    print("Beginning the Prune process")
-    pruned_coeffs, pruned_count_temp = prune_coeffs(
-        coeffs=coeffs, threshold=threshold)
-    total_pruned_count += pruned_count_temp
-    print("Making the new weights")
-    new_weights = reconstruct_weights(
-        pruned_coeffs=pruned_coeffs, wavelet=wavelet, original_shape=original_shape)
-    print("Setting the new weights")
-    layer.set_weights([new_weights] + layer.get_weights()[1:])
-    return pruned_count_temp
-
-
-def handle_non_conv2d_layers(model):
-    """
-    Handle layers without weights and return the number of skipped layers.
+    Prunes the weights of a given TensorFlow layer using a wavelet-based approach.
 
     Args:
-        model: The model containing layers to be processed.
+        layer (tf.keras.layers.Layer): TensorFlow layer to be pruned.
+        threshold (float): Pruning threshold to determine the insignificance of weights.
+        wavelet_type (str): Wavelet type used for the DWT.
+        decomp_level (int): Decomposition level for the DWT.
 
     Returns:
-        int: The number of layers skipped.
+        tf.keras.layers.Layer: Layer with pruned weights.
     """
-    skipped_layers = 0
-    non_conv2d_layers = [layer for layer in model.layers if not isinstance(
-        layer, tf.keras.layers.Conv2D)]
-    for layer in non_conv2d_layers:
-        print(f"Layer {layer.name} has no weights and is skipped.")
-        skipped_layers += 1
-    print(f"Total layers skipped: {skipped_layers}")
-    return skipped_layers
+    # Check if the layer has a kernel attribute
+    if hasattr(layer, 'kernel'):
+        # Get the current weights
+        weights = layer.kernel
+
+        # Apply wavelet decomposition and pruning (TensorFlow-specific implementation)
+        pruned_weights = apply_wavelet_decomposition_tf(
+            weights, wavelet_type, decomp_level, threshold)
+
+        # Update the layer's weights with the pruned weights
+        layer.kernel = pruned_weights
+
+    return layer
 
 
-def log_model_changes(wavelet, level, threshold, skipped_layers, original_param_count, total_pruned_count, guid):
-    model_changes = {
-        "Wavelet": wavelet,
-        "Level": level,
-        "Threshold": threshold,
-        "Layers skipped": skipped_layers,
-        "DWT Phase": 'Yes',
-        "Original Parameter count": original_param_count,
-        "Total pruned count": total_pruned_count,
-    }
-    log_to_csv(guid, model_changes)
-
-
-def apply_dwt(weights, wavelet='haar', level=1):
+def apply_wavelet_decomposition_tf(weights, wavelet_type, decomp_level, threshold):
     """
-    Applies the 2D Discrete Wavelet Transform (DWT) to the input weights matrix.
+    TensorFlow-specific implementation of wavelet decomposition and pruning.
 
     Args:
-        weights (np.ndarray): The weights matrix to be transformed.
-        wavelet (str): The type of wavelet to be used for the transform. Default is 'haar'.
-        level (int): The decomposition level for the wavelet transform. Default is 1.
+        weights (Tensor): Weights tensor from a TensorFlow layer.
+        wavelet_type (str): Type of wavelet used.
+        decomp_level (int): Decomposition level for the DWT.
+        threshold (float): Pruning threshold.
 
     Returns:
-        tuple: A tuple containing the wavelet coefficients and the original shape of the weights matrix.
+        Tensor: Pruned weights tensor.
     """
-    # Handle weights with more than 2 dimensions by reshaping
-    original_shape = weights.shape
-    if len(original_shape) > 2:
-        weights = weights.reshape((-1, original_shape[-1]))
+    # Placeholder for TensorFlow-compatible DWT operations
+    pruned_weights = tf.identity(weights)
+    # Reshape weights to 1D array
+    weights_1d = weights.ravel()
 
-    coeffs = pywt.wavedec2(weights, wavelet, level=level)
+    # Perform wavelet decomposition
+    coeffs = pywt.wavedec(weights_1d, wavelet_type,
+                          mode='periodic', level=decomp_level)
 
-    # Coefficients need to be handled depending on further processing
-    return coeffs, original_shape
-
-
-def prune_coeffs(coeffs, threshold=0.85):
-    """
-    Prunes the wavelet coefficients based on the given threshold and counts the number of coefficients pruned.
-
-    Args:
-        coeffs (list): A list of wavelet coefficients, where each element can be an ndarray or a tuple of ndarrays.
-        threshold (float): The threshold value for pruning coefficients.
-
-    Returns:
-        tuple: A tuple containing:
-            - list: A list of pruned wavelet coefficients.
-            - int: The total number of coefficients pruned.
-    """
+    # Apply thresholding to prune insignificant coefficients
     pruned_coeffs = []
-    total_pruned = 0  # Initialize the count of pruned coefficients
-
     for coeff in coeffs:
-        if isinstance(coeff, tuple):
-            # Initialize a list to hold pruned coefficients for this level
-            pruned_level = []
-            for c in coeff:
-                # Count before pruning
-                pre_prune_count = np.count_nonzero(c)
+        pruned_coeff = pywt.threshold(coeff, threshold, mode='soft')
+        pruned_coeffs.append(pruned_coeff)
 
-                # Prune coefficients
-                pruned_c = np.where(np.abs(c) < threshold, 0, c)
-                pruned_level.append(pruned_c)
+    # Reconstruct the weights using the pruned coefficients
+    pruned_weights_1d = pywt.waverec(
+        pruned_coeffs, wavelet_type, mode='periodic')
 
-                # Count after pruning and update total_pruned
-                post_prune_count = np.count_nonzero(pruned_c)
-                total_pruned += (pre_prune_count - post_prune_count)
+    # Reshape the reconstructed weights back to the original shape
+    pruned_weights = pruned_weights_1d.reshape(weights.shape)
 
-            pruned_coeffs.append(tuple(pruned_level))
-        else:
-            # Handle the approximation coefficients similarly
-            pre_prune_count = np.count_nonzero(coeff)
-            pruned_c = np.where(np.abs(coeff) < threshold, 0, coeff)
-            pruned_coeffs.append(pruned_c)
-            post_prune_count = np.count_nonzero(pruned_c)
-            total_pruned += (pre_prune_count - post_prune_count)
-
-    return pruned_coeffs, total_pruned
-
-# DWT random Weights
+    return pruned_weights
 
 
-def random_pruning(prune_count, guid, model):
+def execute_pruning_workflow(model, threshold, wavelet_type, decomp_level, save_path):
     """
-    prunes randomly.
+    
     """
-    random_pruned_modle = DWTR.randomly_prune_model(
-        model=model, num_prune=prune_count)
-    # model=get_model(FLAGS.model_path), num_prune=prune_count)
-    save_model(random_pruned_modle, original_model_path=FLAGS.model_path,
-               guid=guid, isRandomPruned=True)
+    pruned_weights_info = {}
 
-    # Log model changes
-    model_changes = {
-        "Random prune completed": True,
-    }
-    log_to_csv(guid, model_changes)
+    # Iterate over all layers in the model
+    for layer in model.layers:
+        if hasattr(layer, 'kernel'):
+            # Prune the layer using wavelet-based pruning (TensorFlow-specific implementation)
+            layer = prune_layer_tf(
+                layer, threshold, wavelet_type, decomp_level)
 
+            # Update the pruned weights information
+            kernel_pruned_weights_info = get_pruned_weights_info(layer.kernel)
+            pruned_weights_info[layer.name] = kernel_pruned_weights_info
 
-# IDWT process
-
-
-def reconstruct_weights(pruned_coeffs, wavelet, original_shape):
-    """
-    Reconstructs the weight matrix from the pruned wavelet coefficients.
-
-    Args:
-        pruned_coeffs (list): A list of pruned wavelet coefficients.
-        wavelet (str): The type of wavelet used for the wavelet transform.
-        original_shape (tuple): The original shape of the weight matrix.
-
-    Returns:
-        np.ndarray: The reconstructed weight matrix.
-    """
-    # Reconstruct the weights from the pruned coefficients
-    reconstructed = pywt.waverec2(pruned_coeffs, wavelet)
-
-    # If original weights were reshaped, reshape back to original
-    if len(original_shape) > 2:
-        reconstructed = reconstructed.reshape(original_shape)
-
-    return reconstructed
-
-# Main
+    # Save the pruned model and log the pruning results
+    model_save_path = save_and_log_results(
+        model, pruned_weights_info, save_path, 'DWT')
+    logging.info(f"Pruned model saved to: {model_save_path}")
 
 
 def main(argv):
     """
-    runs the show
+    loads model, then has the execute pruning workflow run, later the random pruning.
     """
+    # Load the pre-trained ResNet model
     setup_tensorflow_gpu()
-    # Parameters
-    wavelet_type = FLAGS.wavelet  # Can be 'db1', 'sym2', etc.
-    decomposition_level = FLAGS.level  # Level of wavelet decomposition
-    threshold = FLAGS.threshold  # Pruning threshold
-    guid = FLAGS.guid
-    # if FLAGS.model_path is None:
-
+    model = TFResNetModel.from_pretrained(FLAGS.model_path)
     # Load the configuration if needed
-    config = ResNetConfig.from_pretrained('microsoft/resnet-18')
-    # Load the pre-trained model
-    model = TFResNetForImageClassification.from_pretrained(
-        'microsoft/resnet-18', config=config)
-    prune_count = optimize_model(threshold=threshold, wavelet=wavelet_type,
-                                 level=decomposition_level, guid=guid, model=model)
-    random_pruning(prune_count, guid, model=model)
+    # config = ResNetConfig.from_pretrained('microsoft/resnet-18')
+    # # Load the pre-trained model
+    # model = TFResNetForImageClassification.from_pretrained(
+    #     'microsoft/resnet-18', config=config)
 
-    # else:
-    #     print(FLAGS.model_path)
-    #     # Percentage of weights to prune selectively
-    #     prune_count = optimize_model(threshold=threshold, wavelet=wavelet_type,
-    #                                  level=decomposition_level, guid=guid)
-    #     random_pruning(prune_count, guid)
+    # Execute the pruning workflow
+    execute_pruning_workflow(
+        model, FLAGS.threshold, FLAGS.wavelet_type, FLAGS.decomp_level, FLAGS.save_path)
 
 
 if __name__ == '__main__':
